Update the structure of django_ecommerce app
