+++ authentication/serializers.py
@@ -3,17 +3,14 @@
 This file contains all the serializers for the authentication Application. The
 RegisterSerializers is used to create the user. The UserViewSerializer is used
 to format the user profile data into json format. The AllUserViewSerializer is
-used to format the all the users data into json format. The CategorySerializer
-is used to create the category. The AllCategorySerializer is used to format the
-all the categories data into json format. The ProductSerializer is used to
-create the product and so on.
+used to format the all the users data into json format.
 """
 
 from rest_framework import serializers
 from django.utils.translation import gettext_lazy as _
 from django.contrib.auth.password_validation import validate_password
 from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
-from .models import Cart, Category, Order, OrderItem, Product, User, ProductReview
+from .models import User
 
 
 # pylint: disable=too-few-public-methods
@@ -110,132 +107,3 @@ class CreateTokneSerialzer(TokenObtainPairSerializer):
         data = super().validate(attrs)
         data["user_id"] = self.user.id
         return data
-
-
-class CategoryViewSerializer(serializers.ModelSerializer):
-    """Category View Serializer
-
-    This class is used to format the category data. The category data is
-    returned in a json format including the category id and name.
-    """
-
-    class Meta:
-        """Meta class for the Category View Serializer"""
-
-        model = Category
-        fields = ["id", "name"]
-
-
-class ProductViewSerializer(serializers.ModelSerializer):
-    """Product View Serializer
-
-    This class is used to format the product data. The product data is
-    returned in a json format including the product id, name, price, category,
-    stock, sold and description.
-    """
-
-    class Meta:
-        """Meta class for the ProductViewSerializer"""
-
-        model = Product
-        fields = [
-            "id",
-            "name",
-            "price",
-            "category",
-            "stock",
-            "sold",
-            "description",
-            "image",
-        ]
-
-    def to_representation(self, instance):
-        representation = super().to_representation(instance)
-        representation["category"] = CategoryViewSerializer(
-            instance.category
-        ).data.get("name")
-        return representation
-
-
-class ReviewSerializer(serializers.ModelSerializer):
-    """Review Serializer
-
-    This serializer is used set the data format for Review Model
-    """
-    class Meta:
-        """Meta class for the ProductReview Serializer"""
-        model = ProductReview
-        fields = ['product', 'user', 'review', 'rating']
-        REQUIRED_FIELD = ['product', 'user']
-
-
-
-class CartSerializer(serializers.ModelSerializer):
-    """Cart Serializer
-
-    This class is used to format the cart data. The cart data is returned in a
-    json format including the cart id, customer and products.
-    """
-
-    class Meta:
-        """Meta class for the Cart Serializer"""
-
-        model = Cart
-        fields = ["user", "product", "quantity", "price"]
-
-    def to_representation(self, instance):
-        representation = super().to_representation(instance)
-        representation["product"] = ProductViewSerializer(
-            instance.product
-        ).data
-        representation["user"] = {
-            "id": instance.user.id,
-            "name": instance.user.username,
-        }
-        return representation
-
-
-class OrderSerializer(serializers.ModelSerializer):
-    """Order Serializer
-
-    This class is used to format the order data. The order data is returned in
-    a json format including the order id, user, price, address, phone number
-    and date_created.
-    """
-
-    class Meta:
-        """Meta class for the OrderSerializer"""
-
-        model = Order
-        fields = ["user", "price", "address", "phone", "date_created"]
-
-    def to_representation(self, instance):
-        representation = super().to_representation(instance)
-        representation["user"] = {
-            "id": instance.user.id,
-            "name": instance.user.username,
-        }
-
-        return representation
-
-
-class OrderDetailSerializer(serializers.ModelSerializer):
-    """Place Order Serializer
-
-    This class is used to format the place order data. The place order data is
-    returned in a json format including the order id and products' detalis.
-    """
-
-    class Meta:
-        """Meta class for the OrderDetailSerializer"""
-
-        model = OrderItem
-        fields = ["order", "product", "quantity", "price", "date_created"]
-
-    def to_representation(self, instance):
-        representation = super().to_representation(instance)
-        representation["order"] = OrderSerializer(instance.order).data
-        representation["product"] = ProductViewSerializer(
-            instance.product
-        ).data
-        return representation
