+++ core/admin.py
@@ -1,3 +1,9 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import Product, Category, Order, Cart
+
+
+admin.site.register(Category)
+admin.site.register(Product)
+admin.site.register(Order)
+admin.site.register(Cart)
